+++ backend/.adminjs/entry.js
@@ -1,1 +1,5 @@
 AdminJS.UserComponents = {}
+import Dashboard from '../src/admin/components/dashboard'
+AdminJS.UserComponents.Dashboard = Dashboard
+import JsonShow from '../src/admin/components/json-show'
+AdminJS.UserComponents.JsonShow = JsonShow